% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDaysSupply.R
\name{checkDaysSupply}
\alias{checkDaysSupply}
\title{Check if Days_supply is the same as datediff(drug_exp_start_date,drug_exp_end_date)}
\usage{
checkDaysSupply(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{byConcept}{whether to get result by concept}
}
\value{
a table with the stats of days supply compared to start and end date
}
\description{
Check if Days_supply is the same as datediff(drug_exp_start_date,drug_exp_end_date)
}
