% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrugSig.R
\name{checkDrugSig}
\alias{checkDrugSig}
\title{Check the drug sig field; this is the verbatim instruction for the drug as
written by the provider.}
\usage{
checkDrugSig(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{byConcept}{whether to get result by drug concept}
}
\value{
a table with a summary of the sig values
}
\description{
Check the drug sig field; this is the verbatim instruction for the drug as
written by the provider.
}
