% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIngredientInTable}
\alias{checkIngredientInTable}
\title{Check ingredient is present in given table}
\usage{
checkIngredientInTable(cdm, conceptId, tableName, messageStore)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{conceptId}{ingredient concept id to check}

\item{tableName}{name of the table to check}

\item{messageStore}{checkmate collection}
}
\description{
Check ingredient is present in given table
}
