% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIsIngredient}
\alias{checkIsIngredient}
\title{Check is an ingredient}
\usage{
checkIsIngredient(cdm, conceptId, messageStore)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{conceptId}{ingredient concept id to check}

\item{messageStore}{checkmate collection}
}
\description{
Check is an ingredient
}
