% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkLogical}
\alias{checkLogical}
\title{Check if given object is a boolean.}
\usage{
checkLogical(input, messageStore, null.ok = TRUE)
}
\arguments{
\item{input}{the input}

\item{messageStore}{checkmate collection}

\item{null.ok}{if value null is allowed}
}
\description{
Check if given object is a boolean.
}
