% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkSampleMinCellCount}
\alias{checkSampleMinCellCount}
\title{Check that the sample is bigger than the mincellcount}
\usage{
checkSampleMinCellCount(sampleSize, minCellCount, messageStore)
}
\arguments{
\item{sampleSize}{sample size for sampling}

\item{minCellCount}{minimum cell count below which to obsure results}

\item{messageStore}{checkmate collection}
}
\description{
Check that the sample is bigger than the mincellcount
}
