% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVerbatimEndDate.R
\name{checkVerbatimEndDate}
\alias{checkVerbatimEndDate}
\title{Check the verbatim_end_date field}
\usage{
checkVerbatimEndDate(
  cdm,
  drugRecordsTable = "ingredient_drug_records",
  byConcept = TRUE,
  sampleSize = 10000
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{modified drug exposure table}

\item{byConcept}{whether to get result by drug concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with the stats about the verbatim_end_date
}
\description{
Check the verbatim_end_date field
}
