% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeDBQuery}
\alias{computeDBQuery}
\title{Store the given input in a remote database table. It will be stored either in a
permanent table or a temporary table depending on tablePrefix.}
\usage{
computeDBQuery(table, tablePrefix, tableName, cdm, overwrite = TRUE)
}
\arguments{
\item{table}{the input table}

\item{tablePrefix}{The stem for the permanent tables that will
be created when running the diagnostics. Permanent tables will be created using
this prefix, and any existing tables that start with this will be at risk of
being dropped or overwritten. If NULL, temporary tables will be
used throughout.}

\item{tableName}{the input table}

\item{cdm}{cdm reference object}

\item{overwrite}{if the table should be overwritten (default TRUE).}
}
\value{
reference to the table
}
\description{
Store the given input in a remote database table. It will be stored either in a
permanent table or a temporary table depending on tablePrefix.
}
