% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{createHistogram}
\alias{createHistogram}
\title{create a histogram for one of days_supply, duration, quantity}
\usage{
createHistogram(cdm, drugRecordsTable = "drug_exposure", type)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{type}{specify whether to plot days_supply, duration or quantity}
}
\value{
object containing a histogram
}
\description{
create a histogram for one of days_supply, duration, quantity
}
