% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataFrame2Hist}
\alias{dataFrame2Hist}
\title{Load an object of class histogram from a data.frame}
\usage{
dataFrame2Hist(df)
}
\arguments{
\item{df}{the dataframe}
}
\value{
an object of class histogram
}
\description{
Load an object of class histogram from a data.frame
}
