% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropTables}
\alias{dropTables}
\title{Drop tables from write_schema of a cdm object}
\usage{
dropTables(cdm, name, verbose = FALSE)
}
\arguments{
\item{cdm}{A cdm reference}

\item{name}{A character vector of tables in the cdm's write_schema}

\item{verbose}{Print a message when dropping a table? TRUE or FALSE (default)}
}
\value{
Invisibly returns the cdm object
}
\description{
cdm objects can have zero or more tables stored in a special schema
where the user has write access. This function removes tables from a cdm's
write_schema
}
