% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllCheckOptions}
\alias{getAllCheckOptions}
\title{Get all options that can be passed to the "checks" parameter}
\usage{
getAllCheckOptions()
}
\description{
Get all options that can be passed to the "checks" parameter
}
