% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRecordChecks.R
\name{getDrugMissings}
\alias{getDrugMissings}
\title{Check missings in drug exposure records}
\usage{
getDrugMissings(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drugRecordsTable}

\item{byConcept}{byConcept}
}
\value{
a table with a summary of missing records
}
\description{
Check missings in drug exposure records
}
