% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRecordChecks.R
\name{getDrugRoutes}
\alias{getDrugRoutes}
\title{Get drug exposure route types}
\usage{
getDrugRoutes(
  cdm,
  drugRecordsTable = "drug_exposure",
  byConcept = TRUE,
  sampleSize = 10000
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{modified drug exposure table}

\item{byConcept}{by individual drug Concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with the drug exposure route types
}
\description{
Get drug exposure route types
}
