% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugStrength.R
\name{getDrugStrength}
\alias{getDrugStrength}
\title{Drug strength records for ingredients of interest}
\usage{
getDrugStrength(
  cdm,
  ingredient,
  includedConceptsTable = "ingredient_concepts",
  drugStrengthTable = "drug_strength",
  tablePrefix = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{ingredient}{ingredient concept ID for ingredient of interest}

\item{includedConceptsTable}{table name for the concept ids, names and units}

\item{drugStrengthTable}{table name for drug strength, default "drug_strength"}

\item{tablePrefix}{The stem for the permanent tables that will
be created when running the diagnostics. Permanent tables will be created using
this prefix, and any existing tables that start with this will be at risk of
being dropped or overwritten. If NULL, temporary tables will be
used throughout.}

\item{verbose}{verbose}
}
\value{
a table containing drug strength records
}
\description{
Drug strength records for ingredients of interest
}
