% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRecordChecks.R
\name{getDrugTypes}
\alias{getDrugTypes}
\title{Get drug exposure record types}
\usage{
getDrugTypes(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drugRecordsTable}

\item{byConcept}{byConcept}
}
\value{
a table with the drug exposure record types
}
\description{
Get drug exposure record types
}
