% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDuration}
\alias{getDuration}
\title{Compute the difference in days between 2 variables in a database table.}
\usage{
getDuration(
  cdm,
  tableName = "drug_exposure",
  startDateCol = "drug_exposure_start_date",
  endDateCol = "drug_exposure_end_date",
  colName = "duration"
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{tableName}{the table name}

\item{startDateCol}{the start date column name}

\item{endDateCol}{the end date column name}

\item{colName}{the result column name}
}
\value{
the table with as new column the duration
}
\description{
Compute the difference in days between 2 variables in a database table.
}
