% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEunomiaCdm}
\alias{getEunomiaCdm}
\title{Get an eunomia CDM reference for given ingredient}
\usage{
getEunomiaCdm(ingredientId = 1125315)
}
\arguments{
\item{ingredientId}{The ingredient concept id}
}
\value{
A list of dplyr database table references pointing to CDM tables
}
\description{
Get an eunomia CDM reference for given ingredient
}
\examples{
\dontrun{
library(CDMConnector)
getEunomiaCdm(1125315)
}
}
