% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIngredientsOverview.R
\name{getIngredientOverview}
\alias{getIngredientOverview}
\title{Get a detailed ingredient overview. The record count and patient count
will be returned for an unique combination of data elements.}
\usage{
getIngredientOverview(
  cdm,
  drugRecordsTable = "drug_exposure",
  drugStrengthTable = "drug_strength"
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{drugStrengthTable}{drug strength table}
}
\value{
a table with the stats
}
\description{
Get a detailed ingredient overview. The record count and patient count
will be returned for an unique combination of data elements.
}
