% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIngredientsOverview.R
\name{getIngredientPresence}
\alias{getIngredientPresence}
\title{Get a presence overview for the ingredient. The record count and patient count
will be returned for a bit set.}
\usage{
getIngredientPresence(
  cdm,
  drugRecordsTable = "drug_exposure",
  drugStrengthTable = "drug_strength"
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{drugStrengthTable}{drug strength table}
}
\value{
a table with the bit set
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
