% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hist2DataFrame}
\alias{hist2DataFrame}
\title{Save an object of class histogram to a data.frame}
\usage{
hist2DataFrame(h)
}
\arguments{
\item{h}{a histogram}
}
\value{
a dataframe with the converted values of the histogram
}
\description{
Save an object of class histogram to a data.frame
}
