% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredientDescendantsInDb.R
\name{ingredientDescendantsInDb}
\alias{ingredientDescendantsInDb}
\title{Get the descendants for the given ingredients}
\usage{
ingredientDescendantsInDb(
  cdm,
  ingredient,
  drugRecordsTable = "drug_exposure",
  tablePrefix = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{ingredient}{ingredient concept id for ingredient of interest}

\item{drugRecordsTable}{table name of the drug exposure records, default "drug_exposure"}

\item{tablePrefix}{The stem for the permanent tables that will
be created when running the diagnostics. Permanent tables will be created using
this prefix, and any existing tables that start with this will be at risk of
being dropped or overwritten. If NULL, temporary tables will be
used throughout.}

\item{verbose}{if verbose set to TRUE, the function will output extra messages}
}
\value{
temp table with concepts used
}
\description{
Get the descendants for the given ingredients
}
