% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDataPanel.R
\name{metaDataPanel}
\alias{metaDataPanel}
\title{metaDataPanel}
\description{
Class to view the metadata of a DrugExposureDiagnostics execution.
}
\section{Super class}{
\code{\link[DrugExposureDiagnostics:ShinyModule]{ShinyModule}} -> \code{metaDataPanel}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-metaDataPanel-initialize}{\code{metaDataPanel$new()}}
    \item \href{#method-metaDataPanel-uiBody}{\code{metaDataPanel$uiBody()}}
    \item \href{#method-metaDataPanel-clone}{\code{metaDataPanel$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DrugExposureDiagnostics" data-topic="ShinyModule" data-id="UI"><a href='../../DrugExposureDiagnostics/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DrugExposureDiagnostics" data-topic="ShinyModule" data-id="server"><a href='../../DrugExposureDiagnostics/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DrugExposureDiagnostics" data-topic="ShinyModule" data-id="validate"><a href='../../DrugExposureDiagnostics/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metaDataPanel-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-metaDataPanel-initialize}{}}}
\subsection{\code{metaDataPanel$new()}}{
  Method to handle the back-end.


  Initializer method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{metaDataPanel$new(data, id, title, description, downloadFilename)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{data from the \code{DrugExposureDiagnostics} package.}
      \item{\code{id}}{the unique reference id for the module}
      \item{\code{title}}{panel title}
      \item{\code{description}}{description of data table}
      \item{\code{downloadFilename}}{filename of the downloaded file}
      \item{\code{input}}{(\code{input})\cr
Input from the server function.}
      \item{\code{output}}{(\code{output})\cr
Output from the server function.}
      \item{\code{session}}{(\code{session})\cr
Session from the server function.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{NULL})
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metaDataPanel-uiBody"></a>}}
\if{latex}{\out{\hypertarget{method-metaDataPanel-uiBody}{}}}
\subsection{\code{metaDataPanel$uiBody()}}{
  Method to include a \link[shiny]{tabPanel} to include the body.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{metaDataPanel$uiBody()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{tabItem})
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-metaDataPanel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-metaDataPanel-clone}{}}}
\subsection{\code{metaDataPanel$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{metaDataPanel$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
