% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscureCounts.R
\name{obscureCounts}
\alias{obscureCounts}
\title{Obscure the small number of counts}
\usage{
obscureCounts(table, tableName, minCellCount = 5, substitute = NA)
}
\arguments{
\item{table}{the table as a tibble}

\item{tableName}{the table name}

\item{minCellCount}{the minimum number of counts that will be displayed. If NULL all results will be reported.}

\item{substitute}{the substitute value if values will be obscured}
}
\value{
the input table with results obscured if minCellCount applies
}
\description{
Obscure the small number of counts
}
