% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printDurationAndMessage}
\alias{printDurationAndMessage}
\title{Print duration from start to now and print it as well as new status message}
\usage{
printDurationAndMessage(message, start)
}
\arguments{
\item{message}{the message}

\item{start}{the start time}
}
\value{
the current time
}
\description{
Print duration from start to now and print it as well as new status message
}
