% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDaysSupply.R
\name{summariseDaysSupply}
\alias{summariseDaysSupply}
\title{Create a summary of the days_supply field}
\usage{
summariseDaysSupply(cdm, drugRecordsTable = "drug_exposure")
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{table name containing the drug exposure records}
}
\value{
a table with the days supply summary
}
\description{
Create a summary of the days_supply field
}
