% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseQuantity.R
\name{summariseQuantity}
\alias{summariseQuantity}
\title{Summarise the quantity column of the drug_exposure table}
\usage{
summariseQuantity(
  cdm,
  drugRecordsTable = "ingredient_drug_records",
  byConcept = TRUE,
  sampleSize = sampleSize
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{modified drug exposure table}

\item{byConcept}{whether to get result by drug concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with the summarized quantity result
}
\description{
Summarise the quantity column of the drug_exposure table
}
