% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateChecks}
\alias{validateChecks}
\title{Validate the "checks" parameter}
\usage{
validateChecks(checks, messageStore)
}
\arguments{
\item{checks}{the checks that have been passed to executeChecks}

\item{messageStore}{checkmate collection}
}
\description{
Validate the "checks" parameter
}
