% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{viewResults}
\alias{viewResults}
\title{View the results in the Shiny app}
\usage{
viewResults(
  dataFolder,
  makePublishable = FALSE,
  publishDir = file.path(getwd(), "ResultsExplorer"),
  overwritePublishDir = FALSE,
  launch.browser = FALSE
)
}
\arguments{
\item{dataFolder}{A folder where the exported zip files with the results are stored.
Zip files containing results from multiple databases can be placed in the same
folder.}

\item{makePublishable}{(Optional) copy data files to make app publishable to posit connect/shinyapp.io}

\item{publishDir}{If make publishable is true - the directory that the shiny app is copied to}

\item{overwritePublishDir}{(Optional) If make publishable is true - overwrite the directory for publishing}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
View the results in the Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the diagnostics
}
