% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeFile}
\alias{writeFile}
\title{Write a result to a file on disk.}
\usage{
writeFile(result, resultName, databaseId, dbDir)
}
\arguments{
\item{result}{check result}

\item{resultName}{name of the result}

\item{databaseId}{database identifier}

\item{dbDir}{output directory for current db}
}
\value{
No return value, called for side effects
}
\description{
Write a result to a file on disk.
}
\examples{
\dontrun{
resultList <- list("mtcars" = mtcars)
result <- writeZipToDisk(
  metadata = metadata,
  databaseId = "mtcars",
  outputFolder = here::here()
)
}
}
