% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeIngredientResultToDisk}
\alias{writeIngredientResultToDisk}
\title{Write (ingredient) diagnostics results on disk in given output folder.}
\usage{
writeIngredientResultToDisk(
  resultList,
  databaseId,
  outputFolder,
  clearDBDir = FALSE
)
}
\arguments{
\item{resultList}{named list with results}

\item{databaseId}{database identifier}

\item{outputFolder}{folder to write to}

\item{clearDBDir}{if database directory should be cleared}
}
\value{
No return value, called for side effects
}
\description{
Write (ingredient) diagnostics results on disk in given output folder.
}
\examples{
\dontrun{
resultList <- list("mtcars" = mtcars)
result <- writeIngredientResultToDisk(
  resultList = resultList,
  databaseId = "mtcars",
  outputFolder = here::here()
)
}
}
