% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeZipToDisk}
\alias{writeZipToDisk}
\title{Write (ingredient) diagnostics results on disk in given output folder.}
\usage{
writeZipToDisk(metadata, databaseId, outputFolder, filename = NULL)
}
\arguments{
\item{metadata}{metadata results}

\item{databaseId}{database identifier}

\item{outputFolder}{folder to write to}

\item{filename}{output filename for the zip file}
}
\value{
No return value, called for side effects
}
\description{
Write (ingredient) diagnostics results on disk in given output folder.
}
\examples{
\dontrun{
resultList <- list("mtcars" = mtcars)
result <- writeZipToDisk(
  metadata = metadata,
  databaseId = "mtcars",
  outputFolder = here::here()
)
}
}
