% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrugDose.R
\name{checkDrugDose}
\alias{checkDrugDose}
\title{Get a summary of the daily drug dose}
\usage{
checkDrugDose(
  cdm,
  drugRecordsTable = "ingredient_drug_records",
  drugStrengthTable = "drug_strength",
  byConcept = TRUE,
  sampleSize = sampleSize
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{modified drug exposure table}

\item{drugStrengthTable}{drug strength table}

\item{byConcept}{whether to get result by drug concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with the stats about the daily dose
}
\description{
Get a summary of the daily drug dose
}
