% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrugDose.R
\name{checkDrugDose}
\alias{checkDrugDose}
\title{Get a summary of the daily drug dose}
\usage{
checkDrugDose(cdm, ingredientConceptId, sampleSize = NULL, minCellCount = 5)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{ingredientConceptId}{ingredient}

\item{sampleSize}{Maximum number of records of an ingredient to estimate dose
coverage. If an ingredient has more, a random sample equal to \code{sampleSize}
will be considered. If NULL, all records will be used.}

\item{minCellCount}{minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}
}
\value{
a table with the stats about the daily dose
}
\description{
Get a summary of the daily drug dose
}
