% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrugSig.R
\name{checkDrugSig}
\alias{checkDrugSig}
\title{Check the drug sig field; this is the verbatim instruction for the drug as
written by the provider.}
\usage{
checkDrugSig(
  cdm,
  drugRecordsTable = "ingredient_drug_records",
  byConcept = TRUE,
  sampleSize = 10000
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{a modified version of the drug exposure table, default "ingredient_drug_records"}

\item{byConcept}{whether to get result by drug concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with a summary of the sig values
}
\description{
Check the drug sig field; this is the verbatim instruction for the drug as
written by the provider.
}
