% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposureDiagnostics-package.R
\docType{package}
\name{DrugExposureDiagnostics-package}
\alias{DrugExposureDiagnostics}
\alias{DrugExposureDiagnostics-package}
\title{DrugExposureDiagnostics: Diagnostics for OMOP Common Data Model Drug Records}
\description{
Ingredient specific diagnostics for drug exposure records in the Observational Medical Outcomes Partnership (OMOP) common data model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/DrugExposureDiagnostics/}
  \item \url{https://github.com/darwin-eu/DrugExposureDiagnostics}
  \item Report bugs at \url{https://github.com/darwin-eu/DrugExposureDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Ger Inberg \email{g.inberg@erasmusmc.nl} (\href{https://orcid.org/0000-0001-8993-8748}{ORCID})

Authors:
\itemize{
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
}

Other contributors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID}) [contributor]
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID}) [contributor]
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID}) [contributor]
  \item Theresa Burkard \email{theresa.burkard@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-1313-4473}{ORCID}) [contributor]
  \item Xintong Li \email{xintong.li@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-6872-5804}{ORCID}) [contributor]
  \item Ross Williams \email{r.williams@erasmusmc.nl} (\href{https://orcid.org/0000-0001-7723-417X}{ORCID}) [contributor]
  \item Erasmus MC [copyright holder]
}

}
\keyword{internal}
