% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeChecks.R
\name{executeChecksSingleIngredient}
\alias{executeChecksSingleIngredient}
\title{Execute all checks on Drug Exposure for a single ingredient.}
\usage{
executeChecksSingleIngredient(
  cdm,
  ingredient = 1125315,
  checks = c("missing", "exposureDuration", "type", "route", "sourceConcept",
    "daysSupply", "verbatimEndDate", "dose", "sig", "quantity", "histogram"),
  minCellCount = 5,
  sample = 1e+06,
  tablePrefix = NULL,
  earliestStartDate = "2010-01-01",
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{ingredient}{ingredient, by default: acetaminophen}

\item{checks}{the checks to be executed, by default everything}

\item{minCellCount}{minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{sample}{the number of samples, default 1 million}

\item{tablePrefix}{The stem for the permanent tables that will
be created when running the diagnostics. Permanent tables will be created using
this prefix, and any existing tables that start with this will be at risk of
being dropped or overwritten. If NULL, temporary tables will be
used throughout.}

\item{earliestStartDate}{the earliest date from which a record can be included}

\item{verbose}{verbose, default FALSE}
}
\value{
named list with results
}
\description{
Execute all checks on Drug Exposure for a single ingredient.
}
