% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRecordChecks.R
\name{getDrugRoutes}
\alias{getDrugRoutes}
\title{Get drug exposure route types}
\usage{
getDrugRoutes(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drugRecordsTable}

\item{byConcept}{byConcept}
}
\value{
a table with the drug exposure route types
}
\description{
Get drug exposure route types
}
