% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRecordChecks.R
\name{getDrugSourceConcepts}
\alias{getDrugSourceConcepts}
\title{Check drug exposure source types}
\usage{
getDrugSourceConcepts(
  cdm,
  drugRecordsTable = "drug_exposure",
  byConcept = TRUE
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drugRecordsTable}

\item{byConcept}{byConcept}
}
\value{
a table with the drug source concepts
}
\description{
Check drug exposure source types
}
