% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseChecks.R
\name{summariseChecks}
\alias{summariseChecks}
\title{Create a summary about the diagnostics results}
\usage{
summariseChecks(resultList)
}
\arguments{
\item{resultList}{a list with the diagnostics results}
}
\value{
a table containing the diagnostics summary
}
\description{
Create a summary about the diagnostics results
}
