% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseQuantity.R
\name{summariseQuantity}
\alias{summariseQuantity}
\title{Summarise the quantity column of the drug_exposure table}
\usage{
summariseQuantity(cdm, drugRecordsTable = "drug_exposure", byConcept = TRUE)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{drug exposure table}

\item{byConcept}{whether to get result by concept}
}
\value{
a table with the summarized quantity result
}
\description{
Summarise the quantity column of the drug_exposure table
}
