% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTimeBetween.R
\name{summariseTimeBetween}
\alias{summariseTimeBetween}
\title{Check time in between drug records per person and report the summary}
\usage{
summariseTimeBetween(
  cdm,
  drugRecordsTable = "ingredient_drug_records",
  byConcept = TRUE,
  sampleSize = 10000
)
}
\arguments{
\item{cdm}{CDMConnector reference object}

\item{drugRecordsTable}{a modified version of the drug exposure table, default "ingredient_drug_records"}

\item{byConcept}{whether to get result by drug concept}

\item{sampleSize}{the sample size given in execute checks}
}
\value{
a table with the stats about the time between
}
\description{
Check time in between drug records per person and report the summary
}
